requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data & ML
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
matplotlib==3.9.0

# HTTP clients
httpx==0.27.0
requests==2.32.3

# Database
sqlalchemy==2.0.30

# Frontend
streamlit==1.35.0

# Email
mailjet-rest==1.3.4

# Environment
python-dotenv==1.0.1